added new folder structure and fix import issues
