--- client/src/pages/admin/Categories/Categories.jsx
@@ -1,11 +1,0 @@
-import { CategoriesTable } from "../../../components/Tables/CategoriesTable";
-
-
-
-export default () => {
-  return (
-    <>
-      <CategoriesTable />
-    </>
-  );
-};
