--- client/src/pages/admin/Products/Products.jsx
@@ -1,10 +1,0 @@
-import { ProductsTable } from "../../../components/Tables/ProductsTable";
-
-
-export default () => {
-  return (
-    <>
-    <ProductsTable/>
-    </>
-  );
-};
