+++ client/src/pages/backend/admin/Categories/Categories.jsx
@@ -1,0 +1,11 @@
+import {CategoriesTable} from "../../../../components/backend/Tables/CategoriesTable"
+
+
+
+export default () => {
+  return (
+    <>
+      <CategoriesTable />
+    </>
+  );
+};
