+++ client/src/pages/backend/admin/Products/Products.jsx
@@ -1,0 +1,10 @@
+import {ProductsTable} from "../../../../components/backend/Tables/ProductsTable"
+
+
+export default () => {
+  return (
+    <>
+    <ProductsTable/>
+    </>
+  );
+};
