+++ client/src/pages/common/errors/DefaultError.jsx
