+++ client/src/pages/common/errors/ServerError.jsx
